Moved DimRange out of RangeFacetSetMatcher
