+++ lucene/facet/src/java/org/apache/lucene/facet/facetset/RangeFacetSetMatcher.java
@@ -17,7 +17,6 @@
 package org.apache.lucene.facet.facetset;
 
 import java.util.Arrays;
-import org.apache.lucene.util.NumericUtils;
 
 /**
  * A {@link FacetSetMatcher} which considers a set as a match if all dimensions fall within the
@@ -66,103 +65,5 @@ public class RangeFacetSetMatcher extends FacetSetMatcher {
       throw new IllegalArgumentException("dimRanges cannot be null or empty");
     }
     return dimRanges.length;
-  }
-
-  /** Defines a single range in a FacetSet dimension. */
-  public static class DimRange {
-    /** Inclusive min */
-    public final long min;
-
-    /** Inclusive max */
-    public final long max;
-
-    /**
-     * Creates a LongRange.
-     *
-     * @param min inclusive min value in range
-     * @param max inclusive max value in range
-     */
-    public DimRange(long min, long max) {
-      this.min = min;
-      this.max = max;
-    }
-
-    /**
-     * Creates a {@link DimRange} for the given min and max long values. This method is also
-     * suitable for int values.
-     */
-    public static DimRange fromLongs(
-        long min, boolean minInclusive, long max, boolean maxInclusive) {
-      if (!minInclusive) {
-        if (min != Long.MAX_VALUE) {
-          min++;
-        } else {
-          throw new IllegalArgumentException("Invalid min input: " + min);
-        }
-      }
-
-      if (!maxInclusive) {
-        if (max != Long.MIN_VALUE) {
-          max--;
-        } else {
-          throw new IllegalArgumentException("Invalid max input: " + max);
-        }
-      }
-
-      if (min > max) {
-        throw new IllegalArgumentException(
-            "Minimum cannot be greater than maximum, max=" + max + ", min=" + min);
-      }
-
-      return new DimRange(min, max);
-    }
-
-    /** Creates a {@link DimRange} for the given min and max double values. */
-    public static DimRange fromDoubles(
-        double min, boolean minInclusive, double max, boolean maxInclusive) {
-      if (Double.isNaN(min)) {
-        throw new IllegalArgumentException("min cannot be NaN");
-      }
-      if (!minInclusive) {
-        min = Math.nextUp(min);
-      }
-
-      if (Double.isNaN(max)) {
-        throw new IllegalArgumentException("max cannot be NaN");
-      }
-      if (!maxInclusive) {
-        max = Math.nextDown(max);
-      }
-
-      if (min > max) {
-        throw new IllegalArgumentException("Minimum cannot be greater than maximum");
-      }
-      return new DimRange(
-          NumericUtils.doubleToSortableLong(min), NumericUtils.doubleToSortableLong(max));
-    }
-
-    /** Creates a {@link DimRange} for the given min and max float values. */
-    public static DimRange fromFloats(
-        float min, boolean minInclusive, float max, boolean maxInclusive) {
-      if (Float.isNaN(min)) {
-        throw new IllegalArgumentException("min cannot be NaN");
-      }
-      if (!minInclusive) {
-        min = Math.nextUp(min);
-      }
-
-      if (Float.isNaN(max)) {
-        throw new IllegalArgumentException("max cannot be NaN");
-      }
-      if (!maxInclusive) {
-        max = Math.nextDown(max);
-      }
-
-      if (min > max) {
-        throw new IllegalArgumentException("Minimum cannot be greater than maximum");
-      }
-      return new DimRange(
-          NumericUtils.floatToSortableInt(min), NumericUtils.floatToSortableInt(max));
-    }
   }
 }
